+++ src/main/java/wanted/preonboard/market/config/CustomUserDetailsService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import wanted.preonboard.market.domain.entity.Member;
+import wanted.preonboard.market.domain.member.Member;
 import wanted.preonboard.market.mapper.MemberMapper;
 
 @Service
