+++ src/main/java/wanted/preonboard/market/domain/Product/dto/ProductInsertDto.java
@@ -1,4 +1,4 @@
-package wanted.preonboard.market.domain.dto;
+package wanted.preonboard.market.domain.Product.dto;
 
 import lombok.Data;
