+++ src/main/java/wanted/preonboard/market/domain/Product/dto/ProductUpdateDto.java
@@ -1,8 +1,8 @@
-package wanted.preonboard.market.domain.dto;
+package wanted.preonboard.market.domain.Product.dto;
 
 import lombok.Data;
 import org.openapitools.jackson.nullable.JsonNullable;
-import wanted.preonboard.market.domain.common.ProductState;
+import wanted.preonboard.market.domain.Product.ProductState;
 
 @Data
 public class ProductUpdateDto {
