+++ src/main/java/wanted/preonboard/market/mapper/MemberMapper.java
@@ -1,8 +1,8 @@
 package wanted.preonboard.market.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
-import wanted.preonboard.market.domain.dto.MemberDto;
-import wanted.preonboard.market.domain.entity.Member;
+import wanted.preonboard.market.domain.member.dto.MemberDto;
+import wanted.preonboard.market.domain.member.Member;
 
 @Mapper
 public interface MemberMapper {
