+++ src/main/java/wanted/preonboard/market/service/Member/MemberServiceImpl.java
@@ -2,8 +2,8 @@ package wanted.preonboard.market.service.Member;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import wanted.preonboard.market.domain.dto.MemberDto;
-import wanted.preonboard.market.domain.entity.Member;
+import wanted.preonboard.market.domain.member.dto.MemberDto;
+import wanted.preonboard.market.domain.member.Member;
 import wanted.preonboard.market.mapper.MemberMapper;
 
 @Service
