+++ src/test/java/wanted/preonboard/market/service/Product/ProductServiceImplTest.java
@@ -6,10 +6,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.openapitools.jackson.nullable.JsonNullable;
-import wanted.preonboard.market.domain.common.ProductState;
-import wanted.preonboard.market.domain.dto.ProductInsertDto;
-import wanted.preonboard.market.domain.dto.ProductUpdateDto;
-import wanted.preonboard.market.domain.entity.Product;
+import wanted.preonboard.market.domain.Product.ProductState;
+import wanted.preonboard.market.domain.Product.dto.ProductInsertDto;
+import wanted.preonboard.market.domain.Product.dto.ProductUpdateDto;
+import wanted.preonboard.market.domain.Product.Product;
 import wanted.preonboard.market.mapper.ProductMapper;
 
 import java.util.Collections;
